kernel: wip: improve debugging
Set the vector table right before the kernel address space is changed.
Before, there was a small gap where the elfloader vector table was not
mapped in, and the kernel vector table was not in use yet. Reduce this
gap.

Create an early uart implementation that continues to use the 1:1 uart
mem mapping in the elfloader until the kernel address space is
activated, where we start using UART_PPTR (the kernel VA for the UART
device) instead.

Signed-off-by: Andy Bui <andy.bui@nio.io>