WIP: Hangjun Patch v2
Previously, we passed the memory regions into the kernel and this was
used to create untypeds. This commit uses the physical memory regions as
described by that list to map the _kernel_ window.

Since each kernel VA is offset from the PA, we do not need to handle PA
reservation (so that they don't become untypeds), arch_init_free_mem
eventually does this for us.